requirements.txt
livekit-agents>=1.3.0
livekit-api>=1.0.0
livekit-plugins-openai>=0.5.0
livekit-plugins-deepgram>=0.2.0
livekit-plugins-google>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
SQLAlchemy>=2.0.0
httpx>=0.28.0
langfuse>=3.0.0
